removed s after Chips folder renaming
