+++ src/index.ts
@@ -4,7 +4,7 @@ export * from './Button'
 export * from './Card'
 export * from './CheckBox'
 export * from './CheckBoxGroup'
-export * from './Chips'
+export * from './Chip'
 export * from './Datepicker'
 export * from './Dropdown'
 export * from './Emoji'
